change search identification enpoint

+++ src/main/java/com/salcedoFawcett/services/clientService/domain/repository/CustomerPartyRepository.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 public interface CustomerPartyRepository {
 
     List<CustomerParty> getAll();
-    Optional<CustomerParty> getCustomerById(long customerID);
-    Optional<CustomerParty> getCustomerByIdentification(double identification);
+    Optional<CustomerParty> getCustomerById(int customerID);
+    Optional<CustomerParty> getCustomerByIdentification(long identification);
     CustomerParty save(CustomerParty customer);
     CustomerParty update(CustomerParty customer);
     boolean customerExist(int customerID);
